Add support for custom effect processors.
Moved all text conversion functions outside the main file, into a new
file called "txtfxcore.js". The "alphabetMap" has now changed to an
"effects" array containing objects describing each effect. See comments
in code for details on the structure of these objects.

Effects now can be either alphabet mappings or custom effects, which have
a processor function as a property of the effect object. When text is
processed, custom effects' processor functions will be called with the
text to be processed as a parameter. Again, see comments in the code for
details on this.