Attempt to fix unwanted caching by telegram.
This is in an attempt to permit users to send the same piece of text over and over again to retry randomized effects like Emojify. This would let them get different emojis on a message if they were dissatisfied with the particular combination that came up.

However, doesn't seem to be effective.
The problem appears to be on the client side - the client refuses to send the same text in a query twice within a certain amount of time.

I'll still commit anyway, may just need some time to take effect or something, or maybe someone else has a different solution.