Add commands system and help information
Typing the /help command in any chat will send a very detailed message containing instructions on how to use the bot.

NOTE: If you are using your own implementation of TxtFXBot, it is recommended that you add a /help command through BotFather so users will not have to type it manually.

Also added note to conversion.py that it will be deleted in the next commit, as its purpose has been served and it is no longer usable; the format has changed and I no longer need it as a utility to create new effects.