Delete conversion.py, as it is no longer needed
This script was used when I was first programming the bot. It was for
creating alphabet objects by inputting all the letters in order.
Since the alphabet format has changed and the script isn't really needed
anyway, I've just decided to delete it in this commit.

You can find the script in previous commits if you really need it, but
be aware it does not generate the correct format.